pyproject.toml
[project]
name = "chat-assistant"
description = "AI Chat Assistant"
version = "0.1.0"
readme = "README.md"
requires-python = "==3.12.0"

dependencies = [
    "typer",
    "pyyaml",
    "rich",
    "pydantic-evals>=1.0.1",
    "sentence-transformers>=5.1.2",
    "chromadb>=1.3.0",
    "langchain-text-splitters>=0.3.11",
    "streamlit>=1.51.0",
    "pydantic-ai>=1.9.1",
    "guardrails-ai>=0.6.7",
    "presidio-analyzer>=2.2.360",
    "presidio-anonymizer>=2.2.360",
    "opentelemetry-exporter-otlp>=1.38.0",
    "opentelemetry-sdk>=1.38.0",
    "bs4>=0.0.2",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.7",
    "numpy>=2.3.4",
    "matplotlib>=3.10.7",
    "openai>=1.109.1",
    "torch>=2.9.0",
    "pypdf>=6.4.0",
    "langsmith>=0.4.46",
]